Fix attribute validator
Remove preference for validator class because it does not exist in older Magento versions
Use before save event instead

refs #65098
