+++ etc/di.xml
@@ -1,8 +1,6 @@
 <?xml version="1.0"?>
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
-    <preference for="Magento\Eav\Model\Validator\Attribute\Code"
-                type="Doofinder\Feed\Model\Eav\Validator\Attribute\DoofinderCode"/>
     <type name="Magento\Config\Model\Config">
         <plugin name="doofinderFeedConfig" type="Doofinder\Feed\Plugin\Config" sortOrder="1" />
     </type>
